// src/resolvers/types.ts
export type Maybe<T> = T | undefined;
// ...
export type Film = Node & {
  __typename?: 'Film';
  characterConnection?: Maybe<FilmCharactersConnection>;
  /** The ISO 8601 date format of the time that this resource was created. */
  created?: Maybe<Scalars['String']>;
  /** The name of the director of this film. */
  director?: Maybe<Scalars['String']>;
  /** The ISO 8601 date format of the time that this resource was edited. */
  edited?: Maybe<Scalars['String']>;
  /** The episode number of this film. */
  episodeID?: Maybe<Scalars['Int']>;
  /** The ID of an object */
  id: Scalars['ID'];
  /** The opening paragraphs at the beginning of this film. */
  openingCrawl?: Maybe<Scalars['String']>;
  planetConnection?: Maybe<FilmPlanetsConnection>;
  /** The name(s) of the producer(s) of this film. */
  producers?: Maybe<Array<Maybe<Scalars['String']>>>;
  /** The ISO 8601 date format of film release at original creator country. */
  releaseDate?: Maybe<Scalars['String']>;
  speciesConnection?: Maybe<FilmSpeciesConnection>;
  starshipConnection?: Maybe<FilmStarshipsConnection>;
  /** The title of this film. */
  title?: Maybe<Scalars['String']>;
  vehicleConnection?: Maybe<FilmVehiclesConnection>;
};
// ...
export type FilmCharactersConnection = {
  __typename?: 'FilmCharactersConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<FilmCharactersEdge>>>;
  /**
   * A list of all of the objects returned in the connection. This is a convenience
   * field provided for quickly exploring the API; rather than querying for
   * "{ edges { node } }" when no edge data is needed, this field can be be used
   * instead. Note that when clients like Relay need to fetch the "cursor" field on
   * the edge to enable efficient pagination, this shortcut cannot be used, and the
   * full "{ edges { node } }" version should be used instead.
   */
  nodes?: Maybe<Array<Maybe<Person>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /**
   * A count of the total number of objects in this connection, ignoring pagination.
   * This allows a client to fetch the first five objects by passing "5" as the
   * argument to "first", then fetch the total count so it could display "5 of 83",
   * for example.
   */
  totalCount?: Maybe<Scalars['Int']>;
};

/** An edge in a connection. */
export type FilmCharactersEdge = {
  __typename?: 'FilmCharactersEdge';
  /** A cursor for use in pagination */
  cursor: Scalars['String'];
  /** The item at the end of the edge */
  node?: Maybe<Person>;
};

/** A connection to a list of items. */
export type FilmPlanetsConnection = {
  __typename?: 'FilmPlanetsConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<FilmPlanetsEdge>>>;
  /**
   * A list of all of the objects returned in the connection. This is a convenience
   * field provided for quickly exploring the API; rather than querying for
   * "{ edges { node } }" when no edge data is needed, this field can be be used
   * instead. Note that when clients like Relay need to fetch the "cursor" field on
   * the edge to enable efficient pagination, this shortcut cannot be used, and the
   * full "{ edges { node } }" version should be used instead.
   */
  nodes?: Maybe<Array<Maybe<Planet>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /**
   * A count of the total number of objects in this connection, ignoring pagination.
   * This allows a client to fetch the first five objects by passing "5" as the
   * argument to "first", then fetch the total count so it could display "5 of 83",
   * for example.
   */
  totalCount?: Maybe<Scalars['Int']>;
};

/** An edge in a connection. */
export type FilmPlanetsEdge = {
  __typename?: 'FilmPlanetsEdge';
  /** A cursor for use in pagination */
  cursor: Scalars['String'];
  /** The item at the end of the edge */
  node?: Maybe<Planet>;
};

/** A connection to a list of items. */
export type FilmSpeciesConnection = {
  __typename?: 'FilmSpeciesConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<FilmSpeciesEdge>>>;
  /**
   * A list of all of the objects returned in the connection. This is a convenience
   * field provided for quickly exploring the API; rather than querying for
   * "{ edges { node } }" when no edge data is needed, this field can be be used
   * instead. Note that when clients like Relay need to fetch the "cursor" field on
   * the edge to enable efficient pagination, this shortcut cannot be used, and the
   * full "{ edges { node } }" version should be used instead.
   */
  nodes?: Maybe<Array<Maybe<Species>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /**
   * A count of the total number of objects in this connection, ignoring pagination.
   * This allows a client to fetch the first five objects by passing "5" as the
   * argument to "first", then fetch the total count so it could display "5 of 83",
   * for example.
   */
  totalCount?: Maybe<Scalars['Int']>;
};
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID'];
};
// ...
export type ResolversObject<TObject> = WithIndex<TObject>;
// ...
export type FilmResolvers<ContextType = any, ParentType extends ResolversParentTypes['Film'] = ResolversParentTypes['Film']> = ResolversObject<{
  characterConnection?: Resolver<Maybe<ResolversTypes['FilmCharactersConnection']>, ParentType, ContextType, Partial<FilmCharacterConnectionArgs>>;
  created?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  director?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  edited?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  episodeID?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
  id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  openingCrawl?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  planetConnection?: Resolver<Maybe<ResolversTypes['FilmPlanetsConnection']>, ParentType, ContextType, Partial<FilmPlanetConnectionArgs>>;
  producers?: Resolver<Maybe<Array<Maybe<ResolversTypes['String']>>>, ParentType, ContextType>;
  releaseDate?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  speciesConnection?: Resolver<Maybe<ResolversTypes['FilmSpeciesConnection']>, ParentType, ContextType, Partial<FilmSpeciesConnectionArgs>>;
  starshipConnection?: Resolver<Maybe<ResolversTypes['FilmStarshipsConnection']>, ParentType, ContextType, Partial<FilmStarshipConnectionArgs>>;
  title?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  vehicleConnection?: Resolver<Maybe<ResolversTypes['FilmVehiclesConnection']>, ParentType, ContextType, Partial<FilmVehicleConnectionArgs>>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type FilmCharactersConnectionResolvers<ContextType = any, ParentType extends ResolversParentTypes['FilmCharactersConnection'] = ResolversParentTypes['FilmCharactersConnection']> = ResolversObject<{
  edges?: Resolver<Maybe<Array<Maybe<ResolversTypes['FilmCharactersEdge']>>>, ParentType, ContextType>;
  nodes?: Resolver<Maybe<Array<Maybe<ResolversTypes['Person']>>>, ParentType, ContextType>;
  pageInfo?: Resolver<ResolversTypes['PageInfo'], ParentType, ContextType>;
  totalCount?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type FilmCharactersEdgeResolvers<ContextType = any, ParentType extends ResolversParentTypes['FilmCharactersEdge'] = ResolversParentTypes['FilmCharactersEdge']> = ResolversObject<{
  cursor?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  node?: Resolver<Maybe<ResolversTypes['Person']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type FilmPlanetsConnectionResolvers<ContextType = any, ParentType extends ResolversParentTypes['FilmPlanetsConnection'] = ResolversParentTypes['FilmPlanetsConnection']> = ResolversObject<{
  edges?: Resolver<Maybe<Array<Maybe<ResolversTypes['FilmPlanetsEdge']>>>, ParentType, ContextType>;
  nodes?: Resolver<Maybe<Array<Maybe<ResolversTypes['Planet']>>>, ParentType, ContextType>;
  pageInfo?: Resolver<ResolversTypes['PageInfo'], ParentType, ContextType>;
  totalCount?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type FilmPlanetsEdgeResolvers<ContextType = any, ParentType extends ResolversParentTypes['FilmPlanetsEdge'] = ResolversParentTypes['FilmPlanetsEdge']> = ResolversObject<{
  cursor?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  node?: Resolver<Maybe<ResolversTypes['Planet']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
// ...
export type Resolvers<ContextType = any> = ResolversObject<{
  Film?: FilmResolvers<ContextType>;
  FilmCharactersConnection?: FilmCharactersConnectionResolvers<ContextType>;
  FilmCharactersEdge?: FilmCharactersEdgeResolvers<ContextType>;
  FilmPlanetsConnection?: FilmPlanetsConnectionResolvers<ContextType>;
  FilmPlanetsEdge?: FilmPlanetsEdgeResolvers<ContextType>;
  FilmSpeciesConnection?: FilmSpeciesConnectionResolvers<ContextType>;
  FilmSpeciesEdge?: FilmSpeciesEdgeResolvers<ContextType>;
  FilmStarshipsConnection?: FilmStarshipsConnectionResolvers<ContextType>;
  FilmStarshipsEdge?: FilmStarshipsEdgeResolvers<ContextType>;
  FilmVehiclesConnection?: FilmVehiclesConnectionResolvers<ContextType>;
  FilmVehiclesEdge?: FilmVehiclesEdgeResolvers<ContextType>;
  FilmsConnection?: FilmsConnectionResolvers<ContextType>;
  FilmsEdge?: FilmsEdgeResolvers<ContextType>;
  JSON?: GraphQLScalarType;
  JSONObject?: GraphQLScalarType;
  Node?: NodeResolvers<ContextType>;
  PageInfo?: PageInfoResolvers<ContextType>;
  PeopleConnection?: PeopleConnectionResolvers<ContextType>;
  PeopleEdge?: PeopleEdgeResolvers<ContextType>;
  Person?: PersonResolvers<ContextType>;
  PersonFilmsConnection?: PersonFilmsConnectionResolvers<ContextType>;
  PersonFilmsEdge?: PersonFilmsEdgeResolvers<ContextType>;
  PersonStarshipsConnection?: PersonStarshipsConnectionResolvers<ContextType>;
  PersonStarshipsEdge?: PersonStarshipsEdgeResolvers<ContextType>;
  PersonVehiclesConnection?: PersonVehiclesConnectionResolvers<ContextType>;
  PersonVehiclesEdge?: PersonVehiclesEdgeResolvers<ContextType>;
  Planet?: PlanetResolvers<ContextType>;
  PlanetFilmsConnection?: PlanetFilmsConnectionResolvers<ContextType>;
  PlanetFilmsEdge?: PlanetFilmsEdgeResolvers<ContextType>;
  PlanetResidentsConnection?: PlanetResidentsConnectionResolvers<ContextType>;
  PlanetResidentsEdge?: PlanetResidentsEdgeResolvers<ContextType>;
  PlanetsConnection?: PlanetsConnectionResolvers<ContextType>;
  PlanetsEdge?: PlanetsEdgeResolvers<ContextType>;
  Query?: QueryResolvers<ContextType>;
  Species?: SpeciesResolvers<ContextType>;
  SpeciesConnection?: SpeciesConnectionResolvers<ContextType>;
  SpeciesEdge?: SpeciesEdgeResolvers<ContextType>;
  SpeciesFilmsConnection?: SpeciesFilmsConnectionResolvers<ContextType>;
  SpeciesFilmsEdge?: SpeciesFilmsEdgeResolvers<ContextType>;
  SpeciesPeopleConnection?: SpeciesPeopleConnectionResolvers<ContextType>;
  SpeciesPeopleEdge?: SpeciesPeopleEdgeResolvers<ContextType>;
  Starship?: StarshipResolvers<ContextType>;
  StarshipFilmsConnection?: StarshipFilmsConnectionResolvers<ContextType>;
  StarshipFilmsEdge?: StarshipFilmsEdgeResolvers<ContextType>;
  StarshipPilotsConnection?: StarshipPilotsConnectionResolvers<ContextType>;
  StarshipPilotsEdge?: StarshipPilotsEdgeResolvers<ContextType>;
  StarshipsConnection?: StarshipsConnectionResolvers<ContextType>;
  StarshipsEdge?: StarshipsEdgeResolvers<ContextType>;
  Vehicle?: VehicleResolvers<ContextType>;
  VehicleFilmsConnection?: VehicleFilmsConnectionResolvers<ContextType>;
  VehicleFilmsEdge?: VehicleFilmsEdgeResolvers<ContextType>;
  VehiclePilotsConnection?: VehiclePilotsConnectionResolvers<ContextType>;
  VehiclePilotsEdge?: VehiclePilotsEdgeResolvers<ContextType>;
  VehiclesConnection?: VehiclesConnectionResolvers<ContextType>;
  VehiclesEdge?: VehiclesEdgeResolvers<ContextType>;
}>;